Add files via upload
I updated the qmd file. I did GLM log-odds, odds and probabilistic analysis on genre after EDA. And considered assigning genre hierarchically and changing it to a numerical variable to be evaluated along with other variables. Ultimately, length, budget and genre were considered to be the main predictors of rating